Refactor event creation into a dedicated component
Moved the event creation form and logic from the `Events.vue` file to a new `CreateEvent.vue` component for improved maintainability and code organization. Added validation for event fields using VeeValidate and Zod. Integrated a "Cancel" event handling mechanism to toggle the form's visibility.
